Renamed and moved some files. Split "logViewerTemplate" into two files
