+++ frontend/src/modules/WellLogViewer/settings/components/TemplateTrackSettings/private-components/SortablePlotList.tsx
@@ -9,7 +9,7 @@ import { ColorSet } from "@lib/utils/ColorSet";
 import { arrayMove } from "@lib/utils/arrays";
 import { TemplatePlotConfig } from "@modules/WellLogViewer/types";
 import { CURVE_COLOR_PALETTE } from "@modules/WellLogViewer/utils/logViewerColors";
-import { PLOT_TYPE_OPTIONS, makeTrackPlot } from "@modules/WellLogViewer/utils/logViewerTemplate";
+import { makeTrackPlot } from "@modules/WellLogViewer/utils/logViewerTemplate";
 import { Delete, SwapHoriz, Warning } from "@mui/icons-material";
 import { TemplatePlotTypes } from "@webviz/well-log-viewer/dist/components/WellLogTemplateTypes";
 
@@ -24,6 +24,21 @@ export type SortablePlotListProps = {
     plots: TemplatePlotConfig[];
     onUpdatePlots: (plots: TemplatePlotConfig[]) => void;
 };
+
+type PlotDropdownOption = DropdownOption<TemplatePlotTypes>;
+
+export const PLOT_TYPE_OPTIONS: PlotDropdownOption[] = [
+    { value: "line", label: "Line" },
+    { value: "linestep", label: "Linestep" },
+    { value: "dot", label: "Dot" },
+    { value: "area", label: "Area" },
+    { value: "gradientfill", label: "Gradientfill" },
+    // TODO: Type requires two named curves, ensure the flow for that is good
+    { value: "differential", label: "Differential" },
+
+    // This one is completely different; requires "discrete" metadata
+    // { value: "stacked", label: "Stacked" },
+];
 
 export function SortablePlotList(props: SortablePlotListProps): React.ReactNode {
     const { onUpdatePlots } = props;
