+++ frontend/src/modules/WellLogViewer/utils/logViewerTemplate.ts
@@ -1,10 +1,8 @@
 /**
  * Utilities and constants used for generating well-log-viewer template configs
  */
-import { DropdownOption } from "@lib/components/Dropdown";
 import {
     Template,
-    TemplatePlotScaleTypes,
     TemplatePlotTypes,
     TemplateTrack,
 } from "@webviz/well-log-viewer/dist/components/WellLogTemplateTypes";
@@ -14,32 +12,9 @@ import { v4 } from "uuid";
 import { CURVE_COLOR_PALETTE, DIFF_CURVE_COLORS } from "./logViewerColors";
 import { MAIN_AXIS_CURVE } from "./queryDataTransform";
 
-import { TemplateTrackConfig } from "../types";
 import { TemplatePlotConfig } from "../types";
 
-type PlotDropdownOption = DropdownOption<TemplatePlotTypes>;
-type TemplatePlotScaleOption = DropdownOption<TemplatePlotScaleTypes>;
-
 export const DEFAULT_MAX_VISIBLE_TRACKS = 5;
-
-export const PLOT_SCALE_OPTIONS: TemplatePlotScaleOption[] = [
-    { label: "Linear", value: "linear" },
-    { label: "Logaritmic", value: "log" },
-];
-
-export const PLOT_TYPES = ["line", "linestep", "dot", "area", "gradientfill", "differential", "stacked"];
-export const PLOT_TYPE_OPTIONS: PlotDropdownOption[] = [
-    { value: "line", label: "Line" },
-    { value: "linestep", label: "Linestep" },
-    { value: "dot", label: "Dot" },
-    { value: "area", label: "Area" },
-    { value: "gradientfill", label: "Gradientfill" },
-    // TODO: Type requires two named curves, ensure the flow for that is good
-    { value: "differential", label: "Differential" },
-
-    // This one is completely different; requires "discrete" metadata
-    // { value: "stacked", label: "Stacked" },
-];
 
 export function isCompositePlotType(type: TemplatePlotTypes) {
     return ["differential"].includes(type);
@@ -104,10 +79,13 @@ export function makeTrackPlot(plot: Partial<TemplatePlotConfig>): TemplatePlotCo
     }
 }
 
+// Matches the ones from the TemplatePlotTypes literal
+const AVAILABLE_PLOT_TYPES = ["line", "linestep", "dot", "area", "gradientfill", "differential", "stacked"];
+
 export function isValidPlot(config: Partial<TemplatePlotConfig>): boolean {
     // This is irregardless of plot type
     if (!config.type || !config.name || !config.color) return false;
-    if (!PLOT_TYPES.includes(config.type)) return false;
+    if (!AVAILABLE_PLOT_TYPES.includes(config.type)) return false;
 
     switch (config.type) {
         case "stacked":
@@ -118,40 +96,5 @@ export function isValidPlot(config: Partial<TemplatePlotConfig>): boolean {
             return Boolean(config.colorTable);
         default:
             return true;
-    }
-}
-
-export function transformToTrackConfigs(objs: any[]): TemplateTrackConfig[] {
-    return objs.map(transformToTrackConfig);
-}
-
-function transformToTrackConfig(obj: any): TemplateTrackConfig {
-    // ! Remember to keep this up to date if the config's structure changes
-    const requiredFields = {
-        title: obj.title,
-        plots: obj.plots,
-    };
-
-    const optionalFields = {
-        _id: obj._id,
-        required: obj.required,
-        width: obj.width,
-        scale: obj.scale,
-        domain: obj.domain,
-    };
-
-    validateRequiredFields(requiredFields);
-
-    return {
-        ...optionalFields,
-        _id: optionalFields._id ?? v4(),
-        title: requiredFields.title,
-        plots: requiredFields.plots.map(makeTrackPlot),
-    };
-}
-
-function validateRequiredFields<T>(partialObj: Partial<T>): asserts partialObj is T {
-    if (Object.values(partialObj).some((v) => v === undefined)) {
-        throw new Error("Missing required fields");
     }
 }
